Fix standards compliance bug in Base64: variable name said equals, but comments and code used various other characters; standards mode not used anywhere else, so if this is wrong somehow, please advise (Python 2.3's base64 module likes the committed code).  Fix return type from SimpleFIeldSet.getInt(String,int).
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@10904 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/support/SimpleFieldSet.java
@@ -503,7 +503,7 @@ public class SimpleFieldSet {
 		}
 	}
 
-	public long getInt(String key, int def) {
+	public int getInt(String key, int def) {
 		String s = get(key);
 		if(s == null) return def;
 		try {
